Relax expensive constraint tying edges to single demands.
Previously we had edges-to-single-vertices used to compute which arcs
only ultimately go to one demand. These arcs could be constrained to
be off when the demand is off, which helps tighten the problem up a
bit. However for large graphs computing which arcs these are becomes
very expensive, so probably not worth it.

This commit relaxes the constraint to only exist for the direct
connectors, so a situation like this:

      /-c-\
D -a-*--b--*-e-S
      \_d_/

Would constrain a <= D. Previously we would have a,b,c,d,e <= D.

Potentially other graph preprocessing could cover this, but I think it
involves clique detection which is also hard in general. There are
some comments left for the interested reader.